Add back main run function
